Rename EmployeeHandler back to EmployeeManager
